requirements.txt
PyGithub>=2.1.1
python-dotenv>=1.0.0
requests>=2.31.0
typer>=0.12.0
rich>=13.7.0
tenacity>=8.2.3
diskcache>=5.6.3
pandas>=2.2.0
